requirements/constraints.txt
# Version constraints for pip-installation.
#
# This file doesn't install any packages. It specifies version constraints
# that will be applied if a package is needed.
#
# When pinning something here, please provide an explanation of why.  Ideally,
# link to other information that will help people in the future to remove the
# pin when possible.  Writing an issue against the offending project and
# linking to it here is good.

# TODO: Add constraint explanation
pylint==2.5.3
pycodestyle==2.5.0
django-admin-extra-buttons==1.5.5
